jagdeep_ws/lib/rnaget/rnaget_service/requirements.txt
jsonschema==3.2.0
werkzeug==0.16.1
connexion==2.7.0
SQLAlchemy==1.3.0
tornado==6.0.1
decorator==4.3.2
openapi-core==0.11.0
pyyaml>=4.2b1
h5py==2.9.0
swagger-ui-bundle==0.0.5
pandas==0.24.2
loompy==2.0.17
prometheus-flask-exporter==0.13.0

# jsonschema[format] requires the following otherwise we get an error:
rfc3987==1.3.8
strict-rfc3339==0.7
webcolors==1.8.1
